refactor: Optimize camera initialization in scene1

+++ static/assets/shaders/shader1.frag
@@ -2,7 +2,7 @@
 
 #extension GL_OES_standard_derivatives : enable
 
-precision highp float;
+precision mediump float;
 
 uniform float time;
 uniform vec2 mouse;
@@ -22,4 +22,5 @@ void main( void ) {
 	gl_FragColor = vec4( vec3( color, color * 0.5, sin( color + time / 3.0 ) * 0.75 ), 1.0 );
 
 }
+
